--- alitheia/core/src/test/java/eu/sqooss/impl/service/updater/test/UpdaterJobWrapper.java
@@ -1,16 +1,0 @@
-package eu.sqooss.impl.service.updater.test;
-
-import eu.sqooss.impl.service.updater.UpdaterJob;
-import eu.sqooss.service.updater.MetadataUpdater;
-
-public class UpdaterJobWrapper extends UpdaterJob{
-
-	public UpdaterJobWrapper(MetadataUpdater updater) {
-		super(updater);
-	}
-
-	public void run() throws Exception{
-		super.run();
-	}
-	
-}
